change project name to akka-aws
